Break out stream_tracer.hpp
